Use autoSync() with the sync command
