+++ cmd/util.go
@@ -12,13 +12,14 @@ import (
 	"strings"
 	"time"
 
+	"golang.org/x/oauth2"
+
 	"github.com/briandowns/spinner"
 	"github.com/fatih/color"
 	"github.com/google/go-github/github"
 	"github.com/knqyf263/pet/config"
-	"github.com/knqyf263/pet/snippet"
-
 	"github.com/knqyf263/pet/dialog"
+	"github.com/knqyf263/pet/snippet"
 )
 
 func autoSync(file string) error {
@@ -43,13 +44,55 @@ func autoSync(file string) error {
 	case local.After(remote):
 		return upload()
 	case remote.After(local):
-		return update(gist)
+		return download(gist)
 	default:
 		return nil
 	}
 }
 
-func update(gist *github.Gist) error {
+func upload() (err error) {
+	ctx := context.Background()
+
+	var snippets snippet.Snippets
+	if err := snippets.Load(); err != nil {
+		return err
+	}
+
+	body, err := snippets.ToString()
+	if err != nil {
+		return err
+	}
+
+	client := githubClient()
+	gist := github.Gist{
+		Description: github.String("description"),
+		Public:      github.Bool(config.Conf.Gist.Public),
+		Files: map[github.GistFilename]github.GistFile{
+			github.GistFilename(config.Conf.Gist.FileName): github.GistFile{
+				Content: github.String(body),
+			},
+		},
+	}
+
+	gistID := config.Conf.Gist.GistID
+	if gistID == "" {
+		var retGist *github.Gist
+		retGist, _, err = client.Gists.Create(ctx, &gist)
+		if err != nil {
+			return err
+		}
+		fmt.Printf("Gist ID: %s\n", retGist.GetID())
+	} else {
+		_, _, err = client.Gists.Edit(ctx, gistID, &gist)
+		if err != nil {
+			return err
+		}
+	}
+	fmt.Println("Upload success")
+	return nil
+}
+
+func download(gist *github.Gist) error {
 	var (
 		content     = ""
 		snippetFile = config.Conf.General.SnippetFile
@@ -73,7 +116,7 @@ func update(gist *github.Gist) error {
 		return err
 	}
 	if content == body {
-		// no need to update
+		// no need to download
 		return nil
 	}
 
@@ -147,4 +190,13 @@ func filter(options []string) (commands []string, err error) {
 		commands = append(commands, fmt.Sprint(snippetInfo.Command))
 	}
 	return commands, nil
+}
+
+func githubClient() *github.Client {
+	ts := oauth2.StaticTokenSource(
+		&oauth2.Token{AccessToken: config.Conf.Gist.AccessToken},
+	)
+	tc := oauth2.NewClient(oauth2.NoContext, ts)
+	client := github.NewClient(tc)
+	return client
 }
